Add audio session management and playback functionality to Deepgram module
- Implemented audio session lifecycle management in Deepgram+AudioSession, including activation, deactivation, and configuration for recording and playback.
- Added handling for audio route changes, media services resets, and audio interruptions to ensure seamless audio experience.
- Introduced Deepgram+Playback for TTS and voice agent audio output using AVAudioPlayerNode, with support for echo cancellation when voice processing is enabled.
- Created Deepgram+Recording to manage microphone capture, supporting both AudioQueue and AVAudioEngine paths for different use cases.
- Enhanced private interface in Deepgram+Private.h to manage shared state across audio session, recording, and playback functionalities.
- Added error handling and logging throughout the audio management process to aid in debugging and ensure reliability.
